혜진 - 사업자 회원가입 오류 수정

+++ cheatingday/src/main/java/com/icia/cheatingday/main/controller/rest/MainRestController.java
@@ -17,14 +17,12 @@ public class MainRestController {
 	private MainRestService service;
 
 	// [일반] 아이디 중복확인
-
 	@GetMapping("/main/u_id_check")
 	public ResponseEntity<Boolean> checkUserId(@RequestParam @NotNull String uUsername) {
 		return ResponseEntity.ok(service.checkUserId(uUsername));
 	}
 
 	// [일반] 이메일 중복확인
-
 	@GetMapping("/main/u_email_check")
 	public ResponseEntity<Boolean> checkUserEmail(@RequestParam @NotNull String uEmail) {
 		return ResponseEntity.ok(service.checkUserEmail(uEmail));
@@ -41,16 +39,14 @@ public class MainRestController {
 	}
 
 	// [사업자] 이메일 중복확인
-
 	@GetMapping("/main/m_email_check")
 	public ResponseEntity<Boolean> checkManagerEmail(@RequestParam @NotNull String mEmail) {
 		return ResponseEntity.ok(service.checkManagerEmail(mEmail));
 	}
 
 	// [사업자] 사업자 등록번호 중복확인
-
 	@GetMapping("/main/m_num_check")
-	public ResponseEntity<Boolean> checkManagerNumber(@RequestParam @NotNull int mNum) {
+	public ResponseEntity<Boolean> checkManagerNumber(@RequestParam @NotNull long mNum) {
 		return ResponseEntity.ok(service.checkManagerNum(mNum));
 	}
